I had mistakenly implemented the ModelHistory class so as to contain only one modified date. This is in fact incorrect :-(
So, the ModelHistory class now has a member List of modified dates
and functions addModifiedDate(), getModifiedDate(unisgned int), getNumModifiedDates() and getListModifiedDates().

In order to not impact existing code the original setModifiedDate and getModifiedDate functions now act on the first element of the list. This seems safe since if people are using these functions they can only have 1 date to act on !!

The only annoying inconsistency is that the C implementation has to call the function that indexes into the list ModelHistory_getModifiedDateFromList
as C does not support functions of the same name with different argument lists.

All the new functions are tested as are parsing to/from xml with multiple modified dates.