Added the following missig fix from the trunk branch.
Fixed "SBase::setNotes(const std::string&)" function so that
the XMLNamespaces object of the SBMLDocument object in which 
the SBase object contained  is used when internally converting 
the given XML string into the corresponding XMLNode object.

Previously, the given XML string may not be converted into
the corresponding XMLNode object if the corresponding xmlns
attribute is not set to the given XML string, leading to 
an empty notes.