(Minor functional change)
Modified the behaviour of 

  XMLNode* XMLNode::convertStringToXMLNode(const std::string&, const XMLNamespaces*) 

function as follows:

   - The function doesn't create a dummy root (container) XMLNode (*) if 
     the given XML string has only one top-level element (e.g. "<p>...</p>").

   - The function creates a dummy root XMLNode if the given XML string
     has two or more top-level elements (e.g.  "<p>...</p><p>...</p>" ).
     Each top-level element in the given XML string is converted to 
     an XMLNode and added to the dummy root XMLNode as its child XMLNode.

     ------------------------------------------------------------------------
     (*) The dummy root XMLNode is an empty XMLNode which is created
         by default constructor as follows:

           XMLNode xmlnode = new XMLNode();

         The empty XMLNode is neither "start", "end", nor "text" element and 
         just used as a container XMLNode. The dummy node is just skipped when
         converting an XMLNode tree to a corresponding XML string.
     ------------------------------------------------------------------------


Previously, the function didn't create a dummy root XMLNode if the top-level 
element in the given XML string is "html", "body", "annotation", or "notes",
otherwise the function created a dummy root XMLNode (i.e. a dummy root XMLNode 
could be created even if the given XML string has only one top-level element 
such as "<p>...</p>".), which is a bit complicated and confusing.

The purpose of this modification is to make the function behave more intuitively.