Implemented some contortions to set the correct Matlab mex architecture variables for some combinations of MacOS and 32 vs 64-bit Matlabs.
The gist of the problem is it turns out to be impossible to influence mex
and mexopts.sh's choice of architecture (particularly "maci" versus
"maci64") solely from the command line arguments given to "mex" nor even
using environment variables set before invoking "mex".  The bug is that
"mex" uses its own value of $Arch *before* it checks for overrides given on
the command line.  The version of mexopts.sh provided with MATLAB uses
$Arch, and since you can't change mex's value of $Arch before it calls its
mexopts.sh, you can't change the settings it makes based on what it thinks
the architecture should be.

I simply could not find a way to influence this short of carrying around
either our own version of "mex" or the "mexopts.sh" file.  The approach
implemented here uses the latter.  The mexopts.sh files provided here use
the environment variable $ARCH instead of $Arch for determining the current
architecture, because mex *does* implement the ability to set $ARCH on the
command line.  Thus, we can set the desired value of ARCH on the command
line, and then our mexopts.sh file can use it.