Trying a solution to the problem that sometimes, after the wrapper process ends, the results file is not yet visible due to (we are guessing) file syste buffering or file system stat() call granularity.  The added code makes run() test for the existence of the results file after the wrapper process ends, and if it doesn't exist, it retries repeatedly for a short while longer until a timeout is reached.
This is suboptimal, but we have a tough problem.  We can't be sure that the wrapper process will *ever* write any file, so we can't wait around indefinitely.  We also have no way of knowing when the file will show up (if it does), since we don't control what the wrappers do when they write the results file and any buffering or flushing or whatever they might do as part of that.  We also don't control the properties of the underlying file system.  

I wish there was a better way, but until we find it, let's try this.