the checkConsistency function will now flatten a comp model after it has validated it and perform validation on the flattened document
I added an error that reports that flattening failed

NOTE: the checkConsistency function on the SBMLDocumentPlugin class now has an optional bool argument that defaults to false. This allows the CompFlatteningConverter to turn off the flattening bit of validation. 

NOTE: the part of the validation code that calls the flattening converter also turns off validation from the converter.