Fixed a serious but obscure bug in FBC wherein the FbcModelPlugin and the Objective didn't implement 'connectToChild' and weren't calling it when constructed.  This led to their children objects mistakenly pointing to the wrong document when the document was copied.  This has the potential to crash programs, and as a side detriment, it totally messes up flattening!
If at all possible, we need to find a way to test this more thoroughly when we create packages, and the package code generator needs to also implement it.  The problem is that a call to 'connectToChild' needs to be added to all constructors that have children:  simply implementing 'connectToChild' doesn't even suffice, because the SBase::SBase() constructors don't call connectToChild, and even if they did, it would actually call the SBase version of the function anyway (as it happens, it calls 'connectToParent' on the plugin children, since the child SBase elements (annotations and notes) don't keep track of their parents (they are XMLNodes).

The whole edifice is pretty fragile, but if we add it to the auto-generation code, I think we can keep ahead of it.  From the look of things, it made it in to the Transition and ListOfFunctionTerms in qual, but not the QualModelPlugin, where it probably needs it for when QualModelPlugins are constructed directly (as noted, 'connectToParent' is called on those objects when the parent SBase object is copied, but not when constructed ex nihilo).

Also, FluxBound didn't need connectToChild, as it had no children.