Filling in documentation that was of the form:
/**
 * @return [...]

because this would make the short form of the documentation for the function completely blank.  Some of these functions are actually set 'internal', but I fixed them anyway, in case we ever un-mark them as internal, or in case they get wrapped by something that wants the documentation anyway.

Also fixed a few 'during while' typos.