Implemented the ability to update the test cases over the network.  This uses the new methods in core Util.java that Frank implemented; I also moved in here the methods from MainWindow.java that dealt with unpacking the internal case archive, so that most all things about the case archives are localized here.
My implementation strikes me as disappointingly complex.  The use of network connections and file handling needs separate threads so that the user interface can present progress bars and dialogs with the option to cancel the operation, but decoupling the separate steps and allowing for cancellation led to a messy combination of runnables, callbacks, dialogs, and associated helper code.  I am pretty sure that I made this more complex than it needed to be, and if I were a better programmer or more adept with Java and SWT, I could have done a better job of it.

This still needs lots more comments, but I'm checking it in because this now mostly seems to work.