Sweep through *all* of the .cpp and .c files in the source to change all instances of /**@cond doxygenCOnly */ to /**@cond doxygenIgnored */.  The reason for this is that we are moving all the C documentation to particular classes/structs, and if doxygen is told that the .h function is a @memberof XX, when it find the same function in the .cpp file, it will *only* think that it is supposed to be documented with whatever is next to it.  Since these functions were already expressly set to be ignored by the C++ documentation (with the COnly directive), and we now don't want it for the C documentation either, changing the @cond allows the file documentation to be (properly) empty, and instead all the documentation ends up in the class groups (and there, only once).
There are several C functions that are exposed in the C++ documentation, particularly those involved with parsing infix and ASTNodes.  If those functions were made @memberof's in the .h file, conditionally '@if conly', the same was done for their implementations in the .cpp/.c files.  This allows the C++ documentation to continue to document those functions, but keeps them in their groups for the C documentation (in the parser case, all these functions are lumped under the 'ASTNode_t' class).

Even if there were no C functions in a @cond/@endcond block, I still changed them, in case C functions are added later.

Also, several #defines were collectively documented in ASTNode.cpp as 'ASTNodeType predicates', but this was only actually added to the first such #define:  now all of them are documented as "ASTNodeType prediate".