Clean up various memory leaks, especially in the Comp flattening converter.  Thanks to Andy for discovering them!
As part of the change, the conversion properties objects in a wide variety of tests are now stack objects instead of pointers--the vast majority of them were never freed.  A few other things were added to the object stack instead of as pointers, too.

Also, a couple new tests here and there that I've been working on and might as well check in now.