Need a file that serves dual duty as short readme and recent news.
The installer created by WhiteBox Packages has only a single slot for
information; they call it "Read me" but it would be nice to also put a
short "recent changes" or "recent news" somewhere.  My idea is to use
use ReleaseNotes.html as a combination readme and summary of recent
changes.