Add some property settings that may help under OS X.
I'm not 100% sure this was the thing that made it work, but after much
trial and error, I've been once again able to get the application name
to show up in the Dock and the tab-switcher in OS X.  Going to leave
this property setting here because it can't hurt.

Note: needs to be in this location, before the SWT thread is started.