Implemented file change monitoring.
This now watches for changes to the results file for the case being
displayed.  If something or someone changes the .csv file externally
without rerunning the case, it will still notice the changes and
refresh the results and plots.

The current implementation only watches for changes to the results for
the case being displayed; so for instance, if case 00100 is being
shown, a change to the results file will get noticed and the plots and
colors updated, but a change to another case like 00101 will not be
noticed until you click on the case.  This is a compromise because
monitoring everything would require a bigger architectural change.