Implementation of changes recommended by Frank:
1) in editPreferences(), don't test for missing date files.  If the
user is using a test directory of their choosing, it may not have a
date file, and then they will get annoying warnings.  Eliminate the
warnings.

2) loadModel(), only attempt to update test cases with the bundled
cases if the user has enabled the feature to check for updates.

3) checkAndUpdateTestCases(...) is newly rewritten to contain behavior
that was previously split between it and loadModel().  Also tries to
be smarter about conditions that cause user prompts.