1) Remove separate libraries bundled in swt-macos.jar 2) Rename swt.jar
At the time I grabbed the latest swt.jar, I should have renamed it
swt-macos.jar.  That's done now, so that it's parallel to the SWT
libraries for other operating systems.  (This is how we used to have
it, but for some reason when I updated SWT to 4.6.2, I got confused
about their library names.)