Rewrote comparison algorithm and `CompareResultSets` class
This now embodies the functionality of `ResultSet.diff()` and
`ResultSet.diffRow()`, so that a single function can be used to compare
result sets and obtain the array of numerical differences.  I did this
because `ResultSet.diff()` did not take into account the possibility of
different column ordering, while `CompareResultSets.compare()` did.  The
GUI needs to recompute both the match/mismatch state and the numerical
difference array on demand, and having these operations split between
functions led to errors.

I took the opportunity to hopefully make `compare()` be slightly more
flexible with respect to column order mismatches, and I added direct
support for comparing a single row (in the manner of `ResultSet.diff()`
and `ResultSet.diffRow()`) because having two separate functions would
have led to near-duplication of a lot of code -- ultimately bad for
maintenance.

Finally, instead of returning a boolean, `compare()` now returns a new
`Comparison` object that encapsulates both the pass/fail boolean and the
numerical differences.