Reimplemented file watching system.
This now uses the Java 8 WatchService functionality.  This Java service turns
out to be slower on macOS (which is something that is complained about on
forums), but trying it anyway because it seems that for some users, the
previous implementation caused a significant CPU load.  This may be less taxing.