Add flags to make it auto-scale on Windows
As reported by Frank, if the user changes the display magnification on
Windows to 150%, the test runner's GUI labels and text get cut off.
After reading the discussion on this page,

  https://bugs.eclipse.org/bugs/show_bug.cgi?id=421383

I added two command-line flags to tell SWT to automatically rescale:

  -Dswt.autoScale=quarter -Dswt.autoScale.method=nearest

The results are not pretty, but at least the text is readable now.