Change how html files are generated
This now uses a scheme whereby the images and css resources are
embedded inside the html files themselves.  This required a different
pipeline to achieve this.  It also now relies on a node.js program
called `inliner`.

Also, this no longer uses `grip`, and just uses `pandoc` (which was the
previous fallback approach if `grip` was unavailable, but `grip` is
mac-only and another dependency, and so in the interest of
simplification and future maintenance I decided to get rid of it).