Abstracted away the file watching code for different OSes
On macOS, Java does not have a good implementation of the WatchService
(even in Java 9).  The WatchService API does exist, but it has an
excessively long response lag, sometimes taking 10 seconds to detect a
file change.  (This is a known issue that others complain about; see
https:*stackoverflow.com/a/11182515/743730) On other OSes, the Java
WatchService is very fast and efficient.  To provide similar behavior
(or rather, to guard against users complaining that something is wrong
with STS on macOS), I created an interface (FileWatcher.java) and then
used two different class implementations, one for macOS and one for
all other systems.  The FileWatcher.java that uses WatchService is now
in NativeFileWatcher.java, and my original polling-based file watcher
is now back, in PollingFileWatcher.java.