Fix shutdown failure
The lack of a shutdown hook meant that when the program was quit using
control-c, it didn't run the proper steps such as saving the settings
XML file.  This sometimes resulted in a zero-length settings file, and
probably meant that threads were not shut down properly either.  This
fix installs a shutdown hook and changes the shutdown sequence slightly.