Change auto-check-updates default to true rather than false
The first time users install this new version of the Test Runner with
a preference setting for auto checking for new test cases, the default
setting really should be true.  They may decide they don't want checks
to happen and manually turn it off, but the initial default makes more
sense to be turned on.