Removed FOSSA line bc inlinder gets a code 500 from the server
In running `src/test-runner/testsuite-ui/build.xml` for creating
distributions, something goes wrong with how `inliner` (invoked by
`build.xml`) calls on the URL, such that the result is a code 500 from
the fossa.io servers.  This is the only one that fails, and it doesn't
fail if I simply open the file in a browser or even attempt a `wget`
on the fossa.io badge URL.  I'm at a loss as to what could be causing
it, and don't have time to debug further, so I'm just going to remove
the fossa.io line for now.