Use newly adjusted icon
Despite that all the colored squares are truly the same size, the red
square always looked shorter when the logo size is small (e.g., when used
as an icon).  I tried to fix this by playing with the size of the red
square, the color of the red, the spacing between the red and yellow
squares, and the drop shadow placement.

Edits were done using Boxy SVG on a mac.