New test cases for https://github.com/sbmlteam/sbml-test-suite/issues/72
The last couple crash the SBML Test Runner, due to there being two 'Time' columns, one for time and one for the variable named 'Time'.  I'm not entirely sure if there's a great solution for this, other than to say 'use the first column as time, and all other columns as element ids'.  But it's a thing that the SBML community needs to work around anyway, since you can indeed have a variable named 'Time'.