Fix disparate external models in comp.
The libsbml flattener does not allow you to reference an external model definition that's not of the same level/version as yourself.  This meant that *all* tests with external models would fail flattening in their l3v2 versions.  This update adds l3v2 versions of all external models, renames the original external models, and points everything to the correct version.